+++ book-service/src/main/java/kr/mybrary/bookservice/book/persistence/repository/interest/BookInterestRepository.java
@@ -1,4 +1,4 @@
-package kr.mybrary.bookservice.book.persistence.repository;
+package kr.mybrary.bookservice.book.persistence.repository.interest;
 
 import java.util.Optional;
 import kr.mybrary.bookservice.book.persistence.Book;
